move n2_node.dart
